backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
razorpay
setuptools
pyjwt
python-dotenv
passlib[bcrypt]
